refactor(collateral): move Conway collateral math to payment-core
The registry builders in `src/services/registry/shared.ts` need the same
declared-total derivation the V2 batch builders already use, but they sit in
the V1-pinned `src/` tree and cannot import from `packages/payment-source-v2`.

Move the math to `@masumi/payment-core/collateral` and re-export it from
`batch-helpers` so every existing batch call site is untouched.

The new module imports no mesh symbol at all, not even a type: `UTxO` is
replaced by a structural `CollateralUtxoLike` that both pinned mesh lines
satisfy. Under ADR 0005 that makes the module safe to share across rails,
because neither mesh line can be pulled in through it.

Behaviour is unchanged. Every moved function and constant is identical to its
original once comments and whitespace are stripped; the only signature change
is `nativeAssetCount(utxo: UTxO)` becoming `CollateralUtxoLike`. Comment
punctuation was normalised in the move, and one paragraph was added to
`computeCollateralFromExUnits` recording that it reads the script half of the
fee only and relies on the safety multiplier and the floor to absorb the size
half.
